+++ app/blockchain/util.js
