+++ app/html/info.html
@@ -6,7 +6,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <meta http-equiv="X-UA-Compatible" content="ie=edge">
     <title>Core App</title>
-    <link rel="stylesheet" href="./styles.css">
+    <link rel="stylesheet" href="../css/info-styles.css">
 </head>
 
 <body>
@@ -20,7 +20,7 @@
     <p id="info"></p>
     <button id="agreeBtn">Ok, got it!</button>
     <script>
-        require("./info.js");
+        require("../js/info.js");
     </script>
 </body>
