+++ app/html/one.html
@@ -5,7 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <meta http-equiv="X-UA-Compatible" content="ie=edge">
     <title>Core App</title>
-    <link rel="stylesheet" href="./styles.css">
+    <link rel="stylesheet" href="../css/one-styles.css">
 </head>
 <body>
     <h1>One</h1>
@@ -18,8 +18,8 @@
     <button id="runBtn">Run</button>
     <p id="status">Ready for search</p>
     <script>
-        require("./hardware-interactions.js");
-        require("./blockchain-interactions.js");
+        require("../js/hardware-interactions.js");
+        require("../js/blockchain-interactions.js");
     </script>
 </body>
 </html>
