+++ app/js/hardware-interactions.js
